backend/requirements.txt
flask
flask-cors
huggingface-hub
python-dotenv